@@ -22,8 +22,8 @@ package vaa.technowize.kandroid.kanboard;
 
 import android.graphics.Color;
 import android.os.AsyncTask;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.util.Log;
 
 import org.json.JSONArray;
