@@ -19,7 +19,7 @@
 
 package vaa.technowize.kandroid.kanboard;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import org.json.JSONObject;
 
