@@ -21,7 +21,7 @@ package vaa.technowize.kandroid.ui;
 
 import android.os.Bundle;
 import android.preference.PreferenceFragment;
-import android.support.v7.app.ActionBar;
+import androidx.appcompat.app.ActionBar;
 import android.view.MenuItem;
 
 import kandroid.R;
